Consolidate typostats coverage gap tests into test_typostats.py
